+++ agrobr/datasets/__init__.py
@@ -2,6 +2,7 @@
 
 from agrobr.datasets.abate_trimestral import abate_trimestral
 from agrobr.datasets.balanco import balanco
+from agrobr.datasets.cadastro_rural import cadastro_rural
 from agrobr.datasets.censo_agropecuario import censo_agropecuario
 from agrobr.datasets.credito_rural import credito_rural
 from agrobr.datasets.custo_producao import custo_producao
@@ -24,6 +25,7 @@ from agrobr.datasets.registry import (
 __all__ = [
     "abate_trimestral",
     "balanco",
+    "cadastro_rural",
     "censo_agropecuario",
     "credito_rural",
     "custo_producao",